TRY_DO_KMODE_TEST
