+++ framework/jinux-frame/src/lib.rs
@@ -14,6 +14,9 @@
 #![feature(generators)]
 #![feature(iter_from_generator)]
 #![feature(const_mut_refs)]
+#![feature(custom_test_frameworks)]
+#![test_runner(jinux_frame::test_runner)]
+#![reexport_test_harness_main = "test_main"]
 
 extern crate alloc;
 
@@ -28,6 +31,7 @@ pub mod logger;
 pub mod prelude;
 pub mod sync;
 pub mod task;
+pub mod test;
 pub mod timer;
 pub mod trap;
 pub mod user;
@@ -37,10 +41,9 @@ pub mod vm;
 pub use self::cpu::CpuLocal;
 pub use self::error::Error;
 pub use self::prelude::Result;
-use alloc::{vec, vec::Vec};
-use core::{mem, panic::PanicInfo};
+use alloc::vec::Vec;
+use core::mem;
 use trap::{IrqCallbackHandle, IrqLine};
-use trapframe::TrapFrame;
 
 static mut IRQ_CALLBACK_LIST: Vec<IrqCallbackHandle> = Vec::new();
 
@@ -61,7 +64,18 @@ pub fn init() {
 fn register_irq_common_callback() {
     unsafe {
         for i in 0..256 {
-            IRQ_CALLBACK_LIST.push(IrqLine::acquire(i as u8).on_active(general_handler))
+            IRQ_CALLBACK_LIST.push(IrqLine::acquire(i as u8).on_active(
+                |_trap_frame| {
+                    todo!()
+                    // info!("general handler");
+                    // println!("{:#x?}", trap_frame);
+                    // println!("rip = 0x{:x}", trap_frame.rip);
+                    // println!("rsp = 0x{:x}", trap_frame.rsp);
+                    // println!("cr2 = 0x{:x}", trap_frame.cr2);
+                    // // println!("rbx = 0x{:x}", trap_frame.)
+                    // panic!("couldn't handle trap right now");
+                }
+            ))
         }
     }
 }
@@ -81,102 +95,7 @@ fn invoke_c_init_funcs() {
     }
 }
 
-fn general_handler(trap_frame: &TrapFrame) {
-    // info!("general handler");
-    // println!("{:#x?}", trap_frame);
-    // println!("rip = 0x{:x}", trap_frame.rip);
-    // println!("rsp = 0x{:x}", trap_frame.rsp);
-    // println!("cr2 = 0x{:x}", trap_frame.cr2);
-    // // println!("rbx = 0x{:x}", trap_frame.)
-    // panic!("couldn't handler trap right now");
-}
-
 #[inline(always)]
 pub(crate) const fn zero<T>() -> T {
     unsafe { mem::MaybeUninit::zeroed().assume_init() }
-}
-
-pub trait Testable {
-    fn run(&self) -> ();
-}
-
-impl<T> Testable for T
-where
-    T: Fn(),
-{
-    fn run(&self) {
-        print!("{}...\n", core::any::type_name::<T>());
-        self();
-        println!("[ok]");
-    }
-}
-
-pub fn test_runner(tests: &[&dyn Testable]) {
-    println!("Running {} tests", tests.len());
-    for test in tests {
-        test.run();
-    }
-    exit_qemu(QemuExitCode::Success);
-}
-
-pub fn test_panic_handler(info: &PanicInfo) -> ! {
-    println!("[failed]");
-    println!("Error: {}", info);
-    exit_qemu(QemuExitCode::Failed);
-}
-
-pub fn panic_handler() {
-    // let mut fp: usize;
-    // let stop = unsafe{
-    //     Task::current().kstack.get_top()
-    // };
-    // info!("stop:{:x}",stop);
-    // unsafe{
-    //     asm!("mov rbp, {}", out(reg) fp);
-    //     info!("fp:{:x}",fp);
-    //     println!("---START BACKTRACE---");
-    //     for i in 0..10 {
-    //         if fp == stop {
-    //             break;
-    //         }
-    //         println!("#{}:ra={:#x}", i, *((fp - 8) as *const usize));
-    //         info!("fp target:{:x}",*((fp ) as *const usize));
-    //         fp = *((fp - 16) as *const usize);
-    //     }
-    //     println!("---END   BACKTRACE---");
-    // }
-}
-
-/// The exit code of x86 QEMU isa debug device. In `qemu-system-x86_64` the
-/// exit code will be `(code << 1) | 1`. So you could never let QEMU invoke
-/// `exit(0)`. We also need to check if the exit code is returned by the
-/// kernel, so we couldn't use 0 as exit_success because this may conflict
-/// with QEMU return value 1, which indicates that QEMU itself fails.
-#[cfg(target_arch = "x86_64")]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u32)]
-pub enum QemuExitCode {
-    Success = 0x10,
-    Failed = 0x20,
-}
-
-pub fn exit_qemu(exit_code: QemuExitCode) -> ! {
-    use x86_64::instructions::port::Port;
-    unsafe {
-        let mut port = Port::new(0xf4);
-        port.write(exit_code as u32);
-    }
-    unreachable!()
-}
-
-#[cfg(feature = "coverage")]
-pub fn get_llvm_coverage_raw() -> Vec<u8> {
-    let mut coverage = vec![];
-    // Safety: minicov::capture_coverage is not thread safe.
-    // There mustn't be any races here.
-    unsafe {
-        minicov::capture_coverage(&mut coverage).unwrap();
-    }
-    minicov::reset_coverage();
-    coverage
 }
